+++ Changes.md
@@ -39,6 +39,7 @@ API
   - Added support in `setValueFromData()` for setting StringPlug values from StringVectorData and StringVectorDataPlugValues from StringData.
   - Added `setValueOrInsertKeyFromData()`.
 - PathListingWidget : Added `visualOrder()`.
+- WidgetAlgo : Added `focusFirstTextWidget()`.
 
 1.5.12.0 (relative to 1.5.11.0)
 ========